Fix undefined behaviour
The `int next_hop = arg.hop - 1;` statements here are relying on C++
undefined behaviour since `arg.hop - 1` is a `uint32_t`: this means in
the case when `arg.hop == 0` that `arg.hop - 1` equals the max uint32_t
value; casting that to an `int` and getting -1 is the undefined
behaviour.

This commit resolves it by just testing that it is strictly positive
in the first place.