Fix net_node logging
LOG_PRINT_L0 is unreliable: its definition depends on the value of
MONERO_DEFAULT_LOG_CATEGORY at the time it is defined, but that is too
late for net_node.inl (which gets included with the header at the end of
net_node.h).

As a result the log file ends up with warnings that include the line
number but *not* the actual message, such as:

    2018-09-05 14:11:12.800 [P2P2]  WARN    net.p2p    src/p2p/net_node.inl:996

This switches them all to MINFO(), which works.

Note that this also deprioritizes the log messages (to info rather than
warning) as these really aren't warnings.