Make has_arg return false for empty vector args
command_line::has_arg(vm, arg) currently always returns true for a
vector arg descriptor because the default, empty vector value is
considered a non-empty value by boost::program_options.

This clearly isn't expected: the code as written expects false in such a
case.  In most cases it didn't actually matter: the `if` guard usually
led to an empty for loop.  But this broke the supernode debugging in
`net_node.inl` quite severely because it ended up trying to dereference
the first element of an empty container and send this to easylogging,
which caused the `net.p2p` debugging to sometimes just produce empty
values.

While it is also possible to simply not do the dereference, it seems a
lot cleaner to simply make `has_arg` do what the code already expects it
to do and return false when called for an argument that has an empty
vector as value.