Use unordered maps/sets; fix cache expiry wrapping
This uses more efficient data structures (unordered_map instead of map,
unordered_set instead of set, and a list instead of a multimap) for
storage supernode-related internals.

It also fixes a bug in the supernode cache that was storing
milliseconds-since-epoch in an `int` to identify the cache expiry, but
this would overflow every 49 days.

More details:

- `set` and `map` are noticeably slower than `unordered_set` and
`unordered_map` whenever you don't need ordering on the values (set) or
keys (map).  Since the request cache and routes are simple lookup
tables, ordering is needed and the unordered versions are a better
choice.

- The request cache expiry structure was being stored as a
std::multimap, but this is entirely unneeded: since entries are always
inserted in order anyway (i.e. new entries are built from the current
time) the overhead of a multimap here is not needed: a simple
std::list<std::pair<...>> does the job just as well (and with
considerably less insert/delete overhead).

- Request cache expiry values were constructed as
milliseconds-since-the-epoch, and then stored in an `int`.  This means
the cache would be severely broken whenever near the wrapping times of
an `int`.

- Optimized cache check/insert/clean into a single method which allows
removing duplicated cache checking/insert/clean code.  It is also
slightly more efficient: we can try the insertion and see if it failed
(resulting it one hash lookup) rather than lookup and then insert
(resulting in two hash lookups).