Add support for building .debs
This builds a shared library version of graft into a set of debs.  The
binary is compiled with Westmere instructions (AES, but not AVX) and
tuned for Haswell.

The resulting binaries/libraries have a `.stable` name appended so that
they can coexist on a system with the `.alpha` version.  The unsuffixed
binaries are controlled via alternatives, but default to this stable
build.