Add support for building .debs -- alpha edition
This builds a shared library version of graft into a set of debs.  The
binary is compiled with Westmere instructions (AES, but not AVX) and
tuned for Haswell.

The resulting binaries/libraries/headers have a `.alpha` name appended
so that they can coexist on a system with the `.stable` version.  The
unsuffixed values are controlled via alternatives, but default to the
stable build when both are installed.